Define `install` and `uninstall` targets
These targets are useful for moving the Gearsystem executable into a
directory located on PATH. These targets aim to adhere to [Makefile
Conventions][1], but have stripped those features that are unnecessary,
given the only output is a single executable.

This has been defined in a separate includable Makefile as these targets
tend to be only applicable to Linux builds. Other distributions may
include this as well if it is applicable.

[1]: <https://www.gnu.org/software/make/manual/make.html#Makefile-Conventions>