Define functional Flatpak App manifest
This manifest allows for the Gearsystem application to start and execute
ROMs without hanging or being unable to render windows.

The `--share=ipc`, `--socket=x11` and `--device=dri` arguments along
with the GLEW and GLU shared modules are necessary for the GUI to appear
whereas the `--socket=pulseaudio` and SDL2 shared module are necessary
for audio to be processed without hitting an indefinite lock.

Filesystem access is still in development, meaning that the
`--filesystem=home` or some equivalent command is required to find ROMs
to load into Gearsystem.