Use NFDe v1.1.1 for Flatpak
Later versions of Native File Dialog Extended (NFDe) have support for
the Flatpak Portal APIs, allowing for native file chooser dialogs that
grant permission to the chosen file without requiring a wide-range of
permissions. This does not allow for SAV file discovery or saving,
however. This will be addressed later.

This is implemented as a separate module—compiled to a static library
rather than as additional source files—to minimise additional
dependencies that would otherwise be handled in its CMake configuration.
As a result, the source files under `nfd/` are not required for the
Flatpak platform and are excluded. Given the Portal API does not use
GTK3 either, the dependencies on this library have also be removed.

A CMake option of `-DCMAKE_INSTALL_INCLUDEDIR=include/nfd` is required
due to the application expecting headers for NFDe to be in a
subdirectory of `nfd/`. This is to mimic the existing include structure,
though it may be argued that keeping them at the root may be better
practice.

This commit necessitates a specific Makefile for the Flatpak platform
given it requires different libraries to link.