Create `.desktop` entry for Flatpak application
This allows the Flatpak application to be added to a desktop environment
such as via a menu item. This is automatically added and removed as the
Flatpak application is installed and uninstalled respectively.

The method of installation was based off of the `.desktop` entry defined
for [Visual Studio Code's Flatpak app][1].

There are currently no icons associated with this application, so it
will appear blank for now.

[1]: https://github.com/flathub/com.visualstudio.code/blob/6ace278a5768a266188e18bc8864ff643cfe22e3/com.visualstudio.code.yaml#L57