Add icons for Flatpak application
This adds a series of PNG files to represent the Gearsystem application.
These adhere to [the Freedesktop Icon specification][1] and come in a
variety of sizes. The icons are based on the existing icon used for the
MacOS release.

Notably, this icon is not used in the panel/taskbar/menubar, but this is
to be investigated.

[1]: https://specifications.freedesktop.org/icon-theme-spec/icon-theme-spec-latest.html