Add MetaInfo file
This file describes the metadata of the project. This can be used on a
variety of repositories, but is currently just used for the Flatpak
application for the time being.

I currently have this MetaInfo file distributed under the "CC0 1.0
Universal" license, essentially putting it in the public domain. This
only covers the metadata file itself and does not impact the license of
the project.

I would also like to mention in this commit that the prior commit does
not have issues with icons. It was instead a local configuration issue.