Make use of containing ROM folder optional
This is mostly useful for distributions where it is not expected for
permission to be granted to the directory a ROM is contained in (e.g.
Flatpak or other Portal API use).

To this end, the `ImGui::Combo` API now uses a list rather that a double
null-terminated string as—combined with trailing commas in C++11—this
allows for entries in this list to be easily added and removed.

Currently, no platform uses this, but this will likely be useful for a
contained release.