Add content rating and developer name
These are to pass the linting process for the AppStream MetaInfo file.
For the content rating, I used [an OARS tool][OARS] to generate the
appropriate tag. In particular, I saw this as an application without
internet connectivity and without any advertising or in-app purchases.
This appears to have generated the most basic content rating.

[OARS]: <https://hughsie.github.io/oars/generate.html>